+++ modules/firewall/CHANGELOG.md
@@ -1,3 +1,41 @@
+##2014-12-16 - Supported Release 1.3.0
+###Summary
+
+This release includes a number of bugfixes and features, including fixing `tcp_flags` support, and added support for interface aliases, negation for iniface and outiface, and extra configurability for packages and service names.
+
+####Features
+- Add support for interface aliases (eth0:0) (MODULES-1469)
+- Add negation for iniface, outiface (MODULES-1470)
+- Make package and service names configurable (MODULES-1309)
+
+####Bugfixes
+- Fix test regexes for EL5 (MODULES-1565)
+- Fix `tcp_flags` support for ip6tables (MODULES-556)
+- Don't arbitrarily limit `set_mark` for certain chains
+
+##2014-11-04 - Supported Release 1.2.0
+###Summary
+
+This release has a number of new features and bugfixes, including rule inversion, future parser support, improved EL7 support, and the ability to purge ip6tables rules.
+
+####Features
+- Documentation updates!
+- Test updates!
+- Add ipset support
+- Enable rule inversion
+- Future parser support
+- Improved support for EL7
+- Support netfilter-persistent
+- Add support for statistics module
+- Add support for mac address source rules
+- Add cbt protocol
+
+####Bugfixes
+- Incorrect use of `source => :iptables` in the ip6tables provider was making it impossible to purge ip6tables rules (MODULES-41)
+- Don't require `toports` when `jump => 'REDIRECT'` (MODULES-1086)
+- Don't limit which chains iniface and outiface parameters can be used in
+- Don't fail on rules added with ipsec/strongswan (MODULES-796)
+
 ##2014-07-08 - Supported Release 1.1.3
 ###Summary
 This is a supported release with test coverage enhancements.